breaking: move smart-values into a dedicated sub-package
